# bos-workspace/backend-nestjs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trả về JWT Token"
          },
          "401": {
            "description": "Sai tài khoản hoặc mật khẩu"
          }
        },
        "summary": "Đăng nhập vào hệ thống",
        "tags": [
          "Auth (Xác thực)"
        ]
      }
    },
    "/api/v1/auth/login/select-tenant": {
      "post": {
        "operationId": "AuthController_loginSelectTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Đăng nhập với doanh nghiệp đã chọn",
        "tags": [
          "Auth (Xác thực)"
        ]
      }
    },
    "/api/v1/auth/my-tenants": {
      "get": {
        "operationId": "AuthController_getMyTenants",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy tất cả các doanh nghiệp liên kết với email hiện tại",
        "tags": [
          "Auth (Xác thực)"
        ]
      }
    },
    "/api/v1/auth/switch-tenant": {
      "post": {
        "operationId": "AuthController_switchTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Chuyển đổi ngữ cảnh doanh nghiệp đang làm việc",
        "tags": [
          "Auth (Xác thực)"
        ]
      }
    },
    "/api/v1/auth/register-tenant": {
      "post": {
        "operationId": "AuthController_registerTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Khởi tạo Doanh nghiệp và tài khoản Admin thành công!"
          },
          "409": {
            "description": "Trùng Email hoặc trùng Mã doanh nghiệp"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đăng ký doanh nghiệp mới (Chỉ Super Admin)",
        "tags": [
          "Auth (Xác thực)"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin tài khoản hiện tại",
        "tags": [
          "Auth (Xác thực)"
        ]
      }
    },
...
    "/api/v1/analytics/entities-summary": {
      "get": {
        "operationId": "AnalyticsController_getEntitiesSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thống kê lượng bản ghi (phiếu) theo từng biểu mẫu",
        "tags": [
          "Macro Analytics (Báo cáo Thống kê Vĩ mô)"
        ]
      }
    },
    "/api/v1/analytics/workflows-summary": {
      "get": {
        "operationId": "AnalyticsController_getWorkflowsSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thống kê trạng thái phê duyệt của các quy trình",
        "tags": [
          "Macro Analytics (Báo cáo Thống kê Vĩ mô)"
        ]
      }
    },
    "/api/v1/analytics/spending-by-department": {
      "get": {
        "operationId": "AnalyticsController_getSpendingByDepartment",
        "parameters": [
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "amountField",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Báo cáo vĩ mô: Thống kê tổng ngân sách đề xuất theo từng Phòng ban (Bóc tách JSONB)",
        "tags": [
          "Macro Analytics (Báo cáo Thống kê Vĩ mô)"
        ]
      }
    },
...
    "/api/v1/automation/events": {
      "get": {
        "operationId": "AutomationController_findEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách các sự kiện kích hoạt khả dụng",
        "tags": [
          "Automation Rules (Quy tắc Tự động hóa)"
        ]
      }
    },
    "/api/v1/automation/rules": {
      "post": {
        "operationId": "AutomationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo mới một quy tắc tự động hóa",
        "tags": [
          "Automation Rules (Quy tắc Tự động hóa)"
        ]
      },
      "get": {
        "operationId": "AutomationController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy toàn bộ quy tắc tự động hóa",
        "tags": [
          "Automation Rules (Quy tắc Tự động hóa)"
        ]
      }
    },
    "/api/v1/automation/rules/{id}": {
      "get": {
        "operationId": "AutomationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy chi tiết một quy tắc",
        "tags": [
          "Automation Rules (Quy tắc Tự động hóa)"
        ]
      },
      "patch": {
        "operationId": "AutomationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật quy tắc tự động hóa",
        "tags": [
          "Automation Rules (Quy tắc Tự động hóa)"
        ]
      },
      "delete": {
        "operationId": "AutomationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa quy tắc tự động hóa",
        "tags": [
          "Automation Rules (Quy tắc Tự động hóa)"
        ]
      }
    }
  },
  "info": {
    "title": "BOS Core Engine API",
    "description": "Tài liệu API cho Nền tảng Quản trị Doanh nghiệp (BOS v2.0)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@bos.com",
            "description": "Email đăng nhập"
          },
          "password": {
            "type": "string",
            "example": "Password@123",
            "description": "Mật khẩu"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SelectTenantDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@bos.com",
            "description": "Email đăng nhập"
          },
          "password": {
            "type": "string",
            "example": "Admin@123",
            "description": "Mật khẩu"
          },
          "tenantId": {
            "type": "number",
            "example": 1,
            "description": "ID Doanh nghiệp đã chọn"
          }
        },
        "required": [
          "email",
          "password",
          "tenantId"
        ]
      },
      "SwitchTenantDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "number",
            "example": 1,
            "description": "ID Doanh nghiệp muốn chuyển sang (Bỏ trống để về platform root)"
          }
        },
        "required": [
          "tenantId"
        ]
      },
      "RegisterTenantDto": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "example": "Công ty Cổ phần Vận tải BOS",
            "description": "Tên doanh nghiệp muốn đăng ký"
          },
          "tenantCode": {
            "type": "string",
            "example": "vantai_bos",
            "description": "Mã định danh duy nhất (snake_case viết thường)"
          },
          "adminEmail": {
            "type": "string",
            "example": "admin@vantaibos.com",
            "description": "Email quản trị viên đầu tiên"
          },
          "adminPassword": {
            "type": "string",
            "example": "Admin@123",
            "description": "Mật khẩu quản trị (tối thiểu 6 ký tự)"
          },
          "adminFullName": {
            "type": "string",
            "example": "Nguyễn Văn Admin",
            "description": "Họ và tên Quản trị viên"
          }
        },
        "required": [
          "tenantName",
          "tenantCode",
          "adminEmail",
          "adminPassword",
          "adminFullName"
        ]
      },
...
      "CreateRuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên quy tắc",
            "example": "Tự động chạy quy trình tiếp theo"
          },
          "eventId": {
            "type": "number",
            "description": "ID của sự kiện kích hoạt",
            "example": 1
          },
          "conditions": {
            "type": "object",
            "description": "Cấu hình điều kiện dạng JSON",
            "example": {}
          },
          "actions": {
            "type": "object",
            "description": "Cấu hình danh sách hành động dạng JSON",
            "example": []
          },
          "isActive": {
            "type": "boolean",
            "description": "Trạng thái hoạt động",
            "example": true
          }
        },
        "required": [
          "name",
          "eventId"
        ]
      },
      "UpdateRuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên quy tắc",
            "example": "Tự động chạy quy trình tiếp theo"
          },
          "eventId": {
            "type": "number",
            "description": "ID của sự kiện kích hoạt",
            "example": 1
          },
          "conditions": {
            "type": "object",
            "description": "Cấu hình điều kiện dạng JSON",
            "example": {}
          },
          "actions": {
            "type": "object",
            "description": "Cấu hình danh sách hành động dạng JSON",
            "example": []
          },
          "isActive": {
            "type": "boolean",
            "description": "Trạng thái hoạt động",
            "example": true
          }
        }
      }
    }
  }
}